better mixins for server load packs stuff

+++ common/src/main/java/net/mehvahdjukaar/moonlight/api/util/FilesHelper.java
@@ -1,5 +1,6 @@
 package net.mehvahdjukaar.moonlight.api.util;
 
+import net.mehvahdjukaar.moonlight.core.Moonlight;
 import org.apache.commons.io.FileUtils;
 
 import java.io.IOException;
@@ -8,7 +9,7 @@ import java.nio.file.Path;
 
 public class FilesHelper {
 
-    public static FastCachedWriter fastCacheWriter(){
+    public static FastCachedWriter fastCacheWriter() {
         return new FastCachedWriter();
     }
 
@@ -25,6 +26,7 @@ public class FilesHelper {
 
         try {
             if (Files.isDirectory(path)) {
+                Moonlight.LOGGER.info("Deleting directory: {}", path);
                 // Move to a temporary location for fast removal
                 Path tempPath = path.resolveSibling(
                         path.getFileName() + "_temp_deleting_" + System.currentTimeMillis()
@@ -40,10 +42,12 @@ public class FilesHelper {
                 new Thread(() -> {
                     try {
                         FileUtils.deleteDirectory(tempPath.toFile());
-                    } catch (Exception ignored) {}
+                    } catch (Exception ignored) {
+                    }
                 }).start();
 
-            } else {
+            } else if (Files.exists(path)) {
+                Moonlight.LOGGER.info("Deleting file: {}", path);
                 // Regular file deletion
                 Files.deleteIfExists(path);
             }
